INSERT UPDATE DELETE Aplikasi Restoran

+++ phpsmk/restoran/kategori/delete.php
@@ -1,12 +1,8 @@
-<?php
-
-    require_once "../function.php";
-
-
-    $sql = "DELETE FROM tblkategori WHERE idkategori = $id";
-
-    $result = mysqli_query($koneksi, $sql);
-
-    header("location:http://localhost/phpsmk/restoran/kategori/select.php");
-
+<?php 
+if(isset($_GET['id'])){
+    $id=$_GET['id'];
+    $sql="DELETE FROM tblkategori WHERE idkategori = $id";
+    $db->runSQL($sql);
+    header("location:?f=kategori&m=select");
+}
 ?>
